Move global declarations to shared/lib folder.
