+++ packages/ts-nameof/lib/declarationFileTests.ts
@@ -5,12 +5,6 @@ import tsNameOf = require("ts-nameof");
 import * as tsNameOfEs6 from "ts-nameof";
 import { IsExactType, assert } from "conditional-type-checks";
 
-namespace TestNamespace {
-    export interface TestType {
-        prop: string;
-    }
-}
-
 /* istanbul ignore next */
 function testFunc() {
     tsNameOf.replaceInFiles(["test"]);
@@ -32,42 +26,4 @@ function testFunc() {
     // es6 test
     const es6Result = tsNameOfEs6.replaceInText("file.ts", "");
     console.log(es6Result.replaced);
-
-    class TestClass {
-        prop: string;
-    }
-
-    // nameof tests
-    nameof(TestClass);
-    nameof<TestNamespace.TestType>();
-    nameof<TestClass>(t => t.prop);
-
-    // nameof.full tests
-    const testInstance = new TestClass();
-    nameof.full(testInstance.prop);
-    nameof.full(testInstance.prop, 1);
-    nameof.full<TestNamespace.TestType>();
-    nameof.full<TestNamespace.TestType>(1);
-    nameof.full<TestClass>(t => t.prop);
-    nameof.full<TestClass>(t => t.prop, 1);
-
-    // reference type test
-    const myObj = { test: "" };
-    nameof(myObj);
-    nameof.full(myObj);
-
-    // primitive type test
-    const myStr = "";
-    nameof(myStr);
-    nameof.full(myStr);
-
-    // null test
-    const nullTypedVar = null;
-    nameof(nullTypedVar);
-    nameof.full(nullTypedVar);
-
-    // undefined test
-    const undefinedTypedVar = undefined;
-    nameof(undefinedTypedVar);
-    nameof.full(undefinedTypedVar);
 }
