+++ shared/lib/global.d.ts
